pyproject.toml
[tool.poetry]
name = "retailcrm-integration-service"
version = "0.1.0"
description = "REST API service for integration with RetailCRM with basic operations on customers and orders"
authors = ["BloodCharry <vladvorobei1994@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = "^0.125.0"
httpx = "^0.28.1"
uvicorn = "^0.38.0"
python-dotenv = "^1.2.1"
structlog = "^24.4.0"
pydantic-settings = "^2.12.0"
email-validator = "^2.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-asyncio = "^1.3.0"
ruff = "^0.14.9"
black = "^25.12.0"
mypy = "^1.19.1"
pre-commit = "^4.5.1"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
exclude = "venv"
strict = true
warn_unused_configs = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true

[[tool.mypy.overrides]]
module = ["tests.*"]
disallow_untyped_defs = false
